successfully modularized js
