+++ styles/main.scss
@@ -1,7 +1,7 @@
 // In order for bootstrap to work, you need to import it like below
 // uncomment the import statements and refresh the DOM to see!
-// @import "~@fortawesome/fontawesome-free/css/all.min.css";
-// @import "~bootstrap/scss/bootstrap";
+@import "~@fortawesome/fontawesome-free/css/all.min.css";
+@import "~bootstrap/scss/bootstrap";
 
 body {
   background-color: rgb(31, 31, 31);
@@ -14,7 +14,7 @@ body {
   color: rgb(194, 194, 194);
   background-color: rgb(44, 44, 44);
   text-align: center;
-  
+
 }
 
 #students {
@@ -79,4 +79,4 @@ body {
   .container {
     flex-direction: column;
   }
-} 
+}
