+++ utils/htmlComp.js
@@ -1,0 +1,83 @@
+// / ********** HTML Components  ********** //
+// the basic HMTL structure of app
+import renderToDOM from './utility';
+
+const htmlStructure = () => {
+  const domString = `
+  <div id="header-container" class="header mb-3"></div>
+  <div id="form-container" class="container mb-3 text-center"></div>
+  <div id="filter-container" class="container mb-3"></div>
+  <div id="student-container" class="container d-flex"></div>
+  `;
+  renderToDOM('#app', domString);
+};
+
+const header = () => {
+  const domString = `<div class="container">
+    <h1>Welcome to Hoggy Hogwarts Sorting Hat!</h1>
+    <p>
+      Hmm, difficult. VERY difficult. <br />Plenty of courage, I see.
+      <br />Not a bad mind, either. There's talent, oh yes. <br />And a
+      thirst to prove yourself. <br />But where to put you?
+    </p>
+  </div>`;
+
+  renderToDOM('#header-container', domString);
+};
+
+const startSortingBtn = () => {
+  const domString = '<button type="button" class="btn btn-info" id="start-sorting">Start the Sorting Ceremony!</button>';
+
+  renderToDOM('#form-container', domString);
+};
+
+const studentAreas = () => {
+  const domString = `<div id="students">No Students</div>
+  <div id="voldy">No Death Eaters</div>`;
+
+  renderToDOM('#student-container', domString);
+};
+
+const studentsOnDom = (divId, array, house = 'Hogwarts') => {
+  let domString = '';
+  if (!array.length) {
+    domString += `NO ${house.toUpperCase()} STUDENTS`;
+  }
+
+  array.forEach((student) => {
+    domString += `
+    <div class="card bg-dark text-white">
+      <img src="${
+  divId === '#voldy'
+    ? 'https://carboncostume.com/wordpress/wp-content/uploads/2019/10/deatheater-harrypotter.jpg' : student.crest}"
+          class="card-img" alt="${student.house} crest">
+      <div class="card-img-overlay">
+        <h5 class="card-title">${student.name}</h5>
+        ${
+  divId === '#voldy'
+    ? '<p class="card-text">Death Eater</p>'
+    : ` <p class="card-text">${student.house}</p>
+          <button type="button" id="expel--${student.id}" class="btn btn-danger btn-sm">Expel</button>`
+}
+      </div>
+    </div>
+    `;
+  });
+  renderToDOM(divId, domString);
+};
+
+const filterBtnRow = () => {
+  const domString = `<div class="btn-group" role="group" aria-label="Basic example">
+      <button type="button" id="filter--hufflepuff" class="btn btn-warning btn-sm">Hufflepuff</button>
+      <button type="button" class="btn btn-primary btn-sm" id="filter--ravenclaw">Ravenclaw</button>
+      <button type="button" class="btn btn-success btn-sm" id="filter--slytherin">Slytherin</button>
+      <button type="button" class="btn btn-danger btn-sm" id="filter--gryffindor">Gryffindor</button>
+      <button type="button" class="btn btn-secondary btn-sm" id="filter--all">All</button>
+    </div>`;
+
+  renderToDOM('#filter-container', domString);
+};
+
+export {
+  htmlStructure, header, startSortingBtn, studentAreas, studentsOnDom, filterBtnRow
+};
