+++ utils/sample_data/student.js
@@ -1,0 +1,5 @@
+const students = [];
+const voldysArmy = []; // starts as an empty array
+export {
+  students, voldysArmy
+};
